requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
jsonschema==4.14.0
tensorflow==2.5.0
tensorflow-hub==0.12.0
sentence-transformers==2.0.0
nltk==3.6.2